[webencodings] Update to 0.6.*
Closes: #16212
