--- stubs/webencodings/webencodings/mklabels.pyi
@@ -1,5 +1,0 @@
-from typing import AnyStr
-from urllib.request import Request
-
-def assert_lower(string: AnyStr) -> AnyStr: ...
-def generate(url: str | Request) -> str: ...
