--- stubs/webencodings/webencodings/x_user_defined.pyi
@@ -1,21 +1,0 @@
-import codecs
-from _codecs import _CharMap
-from _typeshed import ReadableBuffer
-from typing import Final
-
-class Codec(codecs.Codec):
-    def encode(self, input: str, errors: str = "strict") -> tuple[bytes, int]: ...
-    def decode(self, input: bytes, errors: str = "strict") -> tuple[str, int]: ...
-
-class IncrementalEncoder(codecs.IncrementalEncoder):
-    def encode(self, input: str, final: bool = False) -> bytes: ...
-
-class IncrementalDecoder(codecs.IncrementalDecoder):
-    def decode(self, input: ReadableBuffer, final: bool = False) -> str: ...
-
-class StreamWriter(Codec, codecs.StreamWriter): ...
-class StreamReader(Codec, codecs.StreamReader): ...
-
-codec_info: Final[codecs.CodecInfo]
-decoding_table: Final[str]
-encoding_table: Final[_CharMap]
